FIXED: CREATE ROUTER RANK E ME

+++ src/controllers/urls.controller.js
@@ -7,14 +7,13 @@ export async function createShortUrl(req, res) {
 
     try {
         const shortUrl = nanoid(8)
-        await db.query(`INSERT INTO urls (url , "userId") VALUES ($1, $2)`, [text.user.url, text.userId])
+        await db.query(`INSERT INTO urls (url , "userId", "shortUrl") VALUES ($1, $2, $3)`, [text.user.url, text.userId, shortUrl])
 
-        const isUrl = await db.query(`SELECT * FROM urls WHERE "url" = $1`, [text.user.url])
+        const isShort = await db.query(`SELECT * FROM urls WHERE "shortUrl" = $1`, [shortUrl])
 
-        await db.query(`INSERT INTO "shortUrls" ("urlId", "shortUrl") VALUES ($1, $2)`, [isUrl.rows[0].id, shortUrl])
+        await db.query(`UPDATE users SET "linksCount" = "linksCount"+1 WHERE id =$1`, [ isShort.rows[0].userId])
 
-        const isShort = await db.query(`SELECT * FROM "shortUrls" WHERE "shortUrl" = $1`, [shortUrl])
-
+        
         res.status(201).send({ id: isShort.rows[0].id, shortUrl: isShort.rows[0].shortUrl })
 
     } catch (error) {
@@ -24,29 +23,11 @@ export async function createShortUrl(req, res) {
 
 export async function printUrls(req, res) {
     const id = req.params.id
-    // const idUse = +id
-
 
     try {
         const isUrl = await db.query(`SELECT * FROM urls WHERE id = $1`, [id])
 
         if (isUrl.rowCount <= 0) { return res.sendStatus(404) }
-
-        //      const count = await db.query(`SELECT * FROM "shortUrls" WHERE "urlId" = $1`, [isUrl.rows[0].id])
-
-
-        // const visit = await db.query(`SELECT SUM ("visitCount") FROM "shortUrls" WHERE id=$1`, [isUrl.rows[0].urlId])
-
-        // console.log(visit.rows[0],"teste da soma")
-        // await db.query(`UPDATE "shortUrls" SET "visitCount"=$1 WHERE id=$2`, [visit.rows[0].sum   ,isUrl.rows[0].urlId])
-
-        // //console.log(visit.rows[0])`)
-        // console.log(visit.rows[0])
-
-
-        //        const visit = await db.query(`UPDATE "shortUrls" SET "visitCount"="visitCount"+1 WHERE id=$1`, [isUrl.rows[0].urlId])
-
-        //console.log(visit.rows[0])
 
         res.status(200).send({ id: isUrl.rows[0].id, shortUrl: isUrl.rows[0].shortUrl, url: isUrl.rows[0].url })
 
@@ -57,26 +38,20 @@ export async function printUrls(req, res) {
 
 export async function openShortUrl(req, res) {
     const reduceUrl = req.params.shortUrl
-    console.log(reduceUrl, "red");
-
+  
     try {
-        const isShortUrl = await db.query(`SELECT * FROM "shortUrls" WHERE "shortUrl"= $1`, [reduceUrl])
-        if (isShortUrl.rowCount <= 0) { return res.sendStatus(404) }
+        const isShortUrl = await db.query(`SELECT * FROM urls WHERE "shortUrl"= $1`, [reduceUrl])
 
-        const isUrl = await db.query(`SELECT * FROM urls WHERE id= $1`, [isShortUrl.rows[0].urlId])
-        //aumenta o visitCount do Link
-        const visitQtt = await db.query('SELECT "visitCount" FROM "shortUrls" WHERE "shortUrl" = $1', [reduceUrl]);
+           if (isShortUrl.rowCount <= 0) { return res.sendStatus(404) }
 
-        await db.query('UPDATE "shortUrls" SET "visitCount" = $1 WHERE "shortUrl" = $2', [visitQtt.rows[0].visitCount + 1, reduceUrl])
+        await db.query(`UPDATE urls SET "visitCount" = $1 WHERE id =$2`, [isShortUrl.rows[0].visitCount + 1, isShortUrl.rows[0].id])
+        await db.query(`UPDATE users SET "visitCount" = $1 WHERE id =$2`, [isShortUrl.rows[0].visitCount + 1, isShortUrl.rows[0].userId])
 
-        
-        res.redirect(isUrl.rows[0].url)
-
+        res.redirect(isShortUrl.rows[0].url)
 
     } catch (error) {
         return res.status(500).send(error.message)
     }
-
 }
 
 export async function destructionUrl(req, res) {
@@ -92,11 +67,15 @@ export async function destructionUrl(req, res) {
 
         const isUser = await db.query(`SELECT * FROM sessions WHERE token = $1`, [token])
 
+        if (isUser.rowCount <= 0) { return res.sendStatus(401) }
+
         if (isUrl.rows[0].userId !== isUser.rows[0].userId) { return res.sendStatus(401) }
 
 
-        await db.query(`DELETE FROM "shortUrls" WHERE "urlId"=$1`, [isUrl.rows[0].id])
         await db.query(`DELETE FROM urls WHERE id=$1`, [id])
+        
+        
+        await db.query(`UPDATE users SET "linksCount" = "linksCount"-1 WHERE id =$1`, [ isUser.rows[0].userId])
 
         return res.sendStatus(204)
